+++ packages/composites/bsol-price/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/bsol-price-adapter
 
+## 2.2.31
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.7.11
+
 ## 2.2.30
 
 ### Patch Changes