+++ packages/composites/outlier-detection/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/outlier-detection-adapter
 
+## 1.2.54
+
+### Patch Changes
+
+- @chainlink/ea@1.3.53
+
+## 1.2.53
+
+### Patch Changes
+
+- @chainlink/ea@1.3.52
+
 ## 1.2.52
 
 ### Patch Changes